% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_data}
\alias{all_data}
\title{Daily Weather Data from Stations across United States}
\format{
A dataframe with 1,134,354 row and 13 columns
\describe{
  \item{WBANNO}{The weather station WBAN number}
  \item{state}{The state where the station is located}
  \item{station_name}{The name of the station}
  \item{LST_DATE}{The date of observation in YYYYMMDD format}
  \item{CRX_VN}{Version number of the data}
  \item{LONGITUDE}{Longitude of the weather station}
  \item{LATITUDE}{Latitude of the weather station}
  \item{T_DAILY_MAX}{Maximum daily temperature in degrees Celsius}
  \item{T_DAILY_MIN}{Minimum daily temperature in degrees Celsius}
.  \item{T_DAILY_MEAN}{Mean daily temperature in degrees Celsius, by (Max + Min) / 2}
  \item{T_DAILY_AVG}{Average daily temperature in degrees Celsius}
  \item{P_DAILY_CALC}{Calculated daily precipitation in millimeters}
  \item{SOLARAD_DAILY}{Daily solar radiation in MJ/m^2}
}
}
\usage{
all_data
}
\description{
A dataset comprising daily weather observations from weather stations in the US.
The dataset includes various meteorological measurements collected over several years.
}
\keyword{datasets}
