% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_usa_grid}
\alias{create_usa_grid}
\title{Create a grid of points within the contiguous USA}
\usage{
create_usa_grid(resolution = 1)
}
\arguments{
\item{resolution}{The resolution in degrees between points in the grid.
Default is 1 degree.}
}
\value{
A dataframe containing two columns, lon and lat, which represent the longitude
        and latitude of each point within the contiguous USA.
}
\description{
This function generates a grid of longitude and latitude points within the geographic
boundaries of the contiguous United States. The resolution of the grid, which determines
the spacing between grid points, can be specified. Points are filtered to ensure they
fall within the actual boundaries of the USA as defined by the 'maps' package.
}
\examples{
# Create a grid with a resolution of 1 degree
usa_grid <- create_usa_grid()
plot(usa_grid)
# Create a higher resolution grid of 0.5 degrees
finer_usa_grid <- create_usa_grid(0.5)
plot(finer_usa_grid)
}
