% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_prediction.R
\name{grid_prediction}
\alias{grid_prediction}
\title{Make weather predictions on USA map grid points}
\usage{
grid_prediction(usa_grid, data, prediction_value)
}
\arguments{
\item{usa_grid}{a list of grid points of USA map with a longitude column and a latitude column}

\item{data}{a cleaned data set with a LONGITUDE column and a LATITUDE column of stations used for prediction}

\item{prediction_value}{a list of statistic of stations that is used for the prediction of grid points}
}
\value{
A data frame or matrix containing the predicted values for each grid point
        in the `usa_grid`. Each row corresponds to a grid point.
}
\description{
This function is used to make different weather predictions on USA map grid points of various resolutions;
the predictions made depend on the statistic the user is interested in
}
\examples{
# estimate the temperature on March 1st, 2023 for USA map grid points with a given resolution
# create an example USA map grid points with resolution of 1
library(dplyr)
usa_grid <- create_usa_grid(resolution = 2)
data2 <- all_data \%>\% filter(LST_DATE == "2023-03-01")
prediction_value <- data2$T_DAILY_MEAN
grid_pred <- grid_prediction(usa_grid, data2, prediction_value)
print(grid_pred)
}
