% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_prediction.R
\name{map_predictions}
\alias{map_predictions}
\title{Create a heat map for gridded weather predictions}
\usage{
map_predictions(grid_prediction, usa_grid)
}
\arguments{
\item{grid_prediction}{weather predictions made for USA map grids.}

\item{usa_grid}{a list of grid points of USA map with a longitude column and a latitude column. 
This should be identical with the one used to make predictions.}
}
\value{
A ggplot object representing a heat map of the weather predictions across the USA map grid points.
        The heat map displays the geographic distribution of predicted temperatures on the USA map contour.
}
\description{
This function utilizes predictions generated for USA maps grid points to make a heat map on the USA map contour.
Note that the length of predictions should match the grid points.
}
\examples{
# create a heatmap for the estamted temperature on March 1st, 2023 for USA map grid points with a given resolution
#create an example USA map grid points with resolution of 1
usa_grid <- create_usa_grid(resolution = 1)
#create an example station data with LONGITUDE and LATITUDE information used for prediction
data <- data.frame(LONGITUDE = c(-158.61, -164.08, -145.35, -148.46, -150.87,-147.51), 
                    LATITUDE = c(59.28, 61.35, 60.47, 70.16, 63.45,64.97))
# create an example list of statistic of the stations used to predict 
prediction_value <- c( -6.9 -13.4  -1.3 -28.4 -10.0 -11.7)
grid_prediction <- grid_prediction(usa_grid, data, prediction_value)
map_predictions(predictions, usa_grid)

}
