% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{timeser}
\alias{timeser}
\title{Extract time series data for a specific weather station}
\usage{
timeser(df, station_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{df}{A dataframe containing weather data.}

\item{station_id}{A string or numeric value representing the station ID.}

\item{start_date}{Optional start date for filtering data. Default is NULL. If provided,
the function filters data recorded on or after this date.}

\item{end_date}{Optional end date for filtering data. Default is NULL. If provided,
the function filters data recorded on or before this date.}
}
\value{
A filtered dataframe containing data from the specified station between the
       optional start and end dates.
}
\description{
This function filters a given dataset to retrieve data from a specific weather
station within an optional specified date range.
}
\examples{
# Example dataset with columns WBANNO and LST_DATE
data <- data.frame(WBANNO = c(1, 1, 2, 2), LST_DATE = as.Date(c("2023-01-01", "2023-01-02", "2023-01-01", "2023-01-02")),
                   Temp = c(20, 21, 19, 18))
# Extract data for station 1
station1_data <- timeser(data, 1)
print(station1_data)
# Extract data for station 1 from 2023-01-02
station1_data_jan2 <- timeser(data, 1, as.Date("2023-01-02"))
print(station1_data_jan2)
}
