% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_prediction.R
\name{map_predictions}
\alias{map_predictions}
\title{Create a heat map for gridded weather predictions}
\usage{
map_predictions(grid_prediction, usa_grid)
}
\arguments{
\item{grid_prediction}{weather predictions made for USA map grids.}

\item{usa_grid}{a list of grid points of USA map with a longitude column and a latitude column.
This should be identical with the one used to make predictions.}
}
\value{
A ggplot object representing a heat map of the weather predictions across the USA map grid points.
        The heat map displays the geographic distribution of predicted temperatures on the USA map contour.
}
\description{
This function utilizes predictions generated for USA maps grid points to make a heat map on the USA map contour.
Note that the length of predictions should match the grid points.
}
\examples{
# create a heatmap for the estimated temperature on March 1st, 2023 for USA map
# grid points with a given resolution
# create an example USA map grid points with resolution of 1
df <- data("all_daily_data")
usa_grid <- create_usa_grid(resolution = 1)
data2 <- df \%>\% filter(LST_DATE == "2023-03-01")
prediction_value <- data2$T_DAILY_MEAN
grid_pred <- grid_prediction(usa_grid, data2, prediction_value)
output <- map_predictions(grid_pred, usa_grid)
print(output)
}
